requirements.txt
faiss-cpu==1.9.0.post1
fastapi[standard]==0.115.6
langchain==0.3.13
langchain-community==0.3.13
langchain-mistralai==0.2.4
pydantic-settings==2.7.0
pypdf==5.1.0
